requirements.txt
streamlit
pillow
tensorflow
keras
requests
numpy
opencv-python